gst: add sdl support and gstreamer-imx support
Use SDL for event handling so we don't require a lot
of UI dependencies for simple embedded demos of AA.

Add gstreamer-imx plugin support to support efficient
decode and scaling of output on i.MX6.

Scale events from a 1024x600 resolution LCD that I have
on my platform to the 800x480 display that the headunit
code advertises to the handset.

Signed-off-by: Matt Porter <mporter@konsulko.com>